Smart Alarm File Upload
The Smart Alarm Clock was programmed in early 2017 for a course project, using Processing 3.3 (REV 0257) and ARDUINO 1.6.5-r5.

The alarm had two components: an arduino, using multiple ultra-sonic rangefinder sensors as inputs and an lcd display as an output, and a UI created using Processing 3, which can be used to configure the alarm using a USB connection. The alarm was intended to replace the use of a snooze button with rangefinders, which would only activate the alarm if the configured region (a bed) was occupied, and only deactivate if the bed was vacant.